Fix intel_gpu_top not found issue for sudo
Two tests on gstreamer encoding/decoding fail because they do not find
properly intel_gpu_top. That is because the tests are run as normal user
and sudo is used to run intel_gpu_top that is not found because of PATH
value. The solution is to inherit the normal user's PATH value.

KIVU-138