WIP: modify GST_PLUGIN path
This should help with the gstreamer jobs to find the vaapi encoder, but
it does not seem to do the trick.

Another problem is vainfo which does not seem to be able to locate the
proper libraries. When running `vainfo`, it's trying to open
/usr/lib/x86_64-linux-gnu/dri/iHD_drv_video.so, then i965_drv_video.so,
and it fails. Passing a $LD_LIBRARY_PATH that points to the right
directory inside the snap does not help.

Because of that, the vainfo Checkbox job fails and all the jobs are
skipped...